% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_std.R
\name{lm_std}
\alias{lm_std}
\title{linear model standard error}
\usage{
lm_std(x, y, dataset = c())
}
\arguments{
\item{x}{\itemize{
\item corresponds to is the "X" vector that we want to fit our linear regression model for,
}}

\item{y}{\itemize{
\item corresponds to is the "Y" vector that we want to fit our linear regression model for
}}

\item{dataset}{\itemize{
\item the dataset that contains the x, y variable, this dataset variable is set as an empty vector by default if the user wishes to input the x, y vector directly.
}}
}
\value{
return the standard error of the linear model fitted for vector x, y.
}
\description{
This function will return the standard error of the linear model fitted for vector x, y. Since we are fitting a linear model, none-numeric vectors will not be accepted, also we would expect the length of these two vectors to match. The input vector can have NA as long as we have enough numeric elements to fit our model.
}
\examples{
lm_std(c(1,2,3), c(4,5,7))
# should return an integer which is the standard error of the linear model
# fitted for (c(4,5,7) ~ c(1,2,3))
lm_std(datateachr::cancer_sample$radius_mean, datateachr::cancer_sample$perimeter_mean)
# should return an integer which is the standard error of the linear model
# fitted for variable (perimeter_mean ~ radius_mean) inside dataset cancer_sample
}
